pyproject.toml
[tool.poetry]
name = "fast-api-app"
version = "0.1.0"
description = "Build CI/CD pipeline for FastAPI application"
authors = ["RonanRibouletPython <newsletterpython13@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
sqlalchemy = "^2.0.35"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.9"
sqlalchemy-utils = "^0.41.2"
alembic = "^1.13.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.12"


[tool.poetry.group.dev.dependencies]
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
